separating GUI dependencies so they can be reused
